fix: do not use application-config in browser env
