+++ packages/application-config/src/index.ts
@@ -1,4 +1,2 @@
 export { default as processConfig } from './process-config';
-export { buildOidcScope } from './build-oidc-scope';
 export * from './types';
-export * from './constants';
